# docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
...
  CombinedResourceUsage:
    description: CombinedResourceUsage defines use of a combined resource
    properties:
      cpu_resource:
        $ref: '#/definitions/ResourceUsage'
      disk_resource:
        $ref: '#/definitions/ResourceUsage'
      organization_id:
        type: string
        x-go-name: OrganizationID
      user_id:
        type: string
        x-go-name: UserID
    type: object
    x-go-package: plexobject.com/formicary/queen/types
...
  Duration:
    description: |-
      A Duration represents the elapsed time between two instants
      as an int64 nanosecond count. The representation limits the
      largest representable duration to approximately 290 years.
    format: int64
    type: integer
    x-go-package: time
...
  JobResource:
    description: |-
      Job Resources can be used for allocating computing resources such as devices, CPUs, memory, connections, licences, etc.
      You can use them as mutex, semaphores or quota system to determine concurrency of jobs.
      For example, a job may need a license key to connect to a third party service, and it may only accept upto
      five connections that can be allocated via resources.
    properties:
      category:
        description: Category can be used to represent grouping of resources
        type: string
        x-go-name: Category
      created_at:
        description: CreatedAt job creation time
        format: date-time
        type: string
        x-go-name: CreatedAt
      description:
        description: Description of resource
        type: string
        x-go-name: Description
      disabled:
        description: Disabled is used to stop further processing of job, and it can
          be used during maintenance, upgrade or debugging.
        type: boolean
        x-go-name: Disabled
      external_id:
        description: ExternalID defines external-id of the resource if exists
        type: string
        x-go-name: ExternalID
      extract_config:
        $ref: '#/definitions/ResourceCriteriaConfig'
      id:
        description: ID defines UUID for primary key
        type: string
        x-go-name: ID
      lease_timeout:
        $ref: '#/definitions/Duration'
      organization_id:
        description: OrganizationID defines org who submitted the job
        type: string
        x-go-name: OrganizationID
      platform:
        description: Platform can be OS platform or target runtime
        type: string
        x-go-name: Platform
      quota:
        description: |-
          Quota can be used to represent mutex (max 1), semaphores (max limit) or other kind of quota.
          Note: mutex/semaphores can only take one resource by quota may take any value
        format: int64
        type: integer
        x-go-name: Quota
      resource_type:
        description: ResourceType defines type of resource such as Device, CPU, Memory
        type: string
        x-go-name: ResourceType
      resource_variables:
        additionalProperties:
          type: object
        description: Following are transient properties
        type: object
        x-go-name: NameValueConfigs
      tags:
        description: Tags can be used as tags for resource matching
        items:
          type: string
        type: array
        x-go-name: Tags
      updated_at:
        description: UpdatedAt job update time
        format: date-time
        type: string
        x-go-name: UpdatedAt
      user_id:
        description: UserID defines user who submitted the job
        type: string
        x-go-name: UserID
      valid_status:
        description: ValidStatus - health status of resource
        type: boolean
        x-go-name: ValidStatus
      value:
        description: Value consumed, e.g. it will be 1 for mutex, semaphore but can
          be higher number for other quota system
        format: int64
        type: integer
        x-go-name: Value
    title: JobResource represents a virtual resource, which can be used to implement
      mutex/semaphores for a job.
    type: object
    x-go-package: plexobject.com/formicary/queen/types
...
  Kind:
    description: Kind defines enum for kind of subscription
    type: string
    x-go-package: plexobject.com/formicary/internal/types
...
  Organization:
    description: It is used multi-tenancy support in the platform.
    properties:
      bundle_id:
        description: BundleID defines package or bundle
        type: string
        x-go-name: BundleID
      created_at:
        description: CreatedAt created time
        format: date-time
        type: string
        x-go-name: CreatedAt
      id:
        description: |-
          gorm.Model
          ID defines UUID for primary key
        type: string
        x-go-name: ID
      license_policy:
        description: LicensePolicy defines license policy
        type: string
        x-go-name: LicensePolicy
      max_concurrency:
        description: MaxConcurrency defines max number of jobs that can be run concurrently
          by org
        format: int64
        type: integer
        x-go-name: MaxConcurrency
      org_unit:
        description: OrgUnit defines org-unit
        type: string
        x-go-name: OrgUnit
      owner_user_id:
        description: OwnerUserID defines owner user
        type: string
        x-go-name: OwnerUserID
      parent_id:
        description: ParentID defines parent org
        type: string
        x-go-name: ParentID
      salt:
        description: Salt for password
        type: string
        x-go-name: Salt
      sticky_message:
        description: StickyMessage defines an error message that needs user attention
        type: string
        x-go-name: StickyMessage
      subscription:
        $ref: '#/definitions/Subscription'
      updated_at:
        description: UpdatedAt update time
        format: date-time
        type: string
        x-go-name: UpdatedAt
    title: Organization represents a user organization that may have one or more users.
    type: object
    x-go-package: plexobject.com/formicary/internal/types
...
  Period:
    description: Period defines enum for period of subscription
    type: string
    x-go-package: plexobject.com/formicary/internal/types
  Policy:
    description: Policy defines enum for policy of subscription
    type: string
    x-go-package: plexobject.com/formicary/internal/types
...
  ResourceCriteriaConfig:
    title: ResourceCriteriaConfig defines properties to extract from a resource into
      job variables.
...
  Subscription:
    description: Subscription defines subscription
    properties:
      active:
        description: Active flag
        type: boolean
        x-go-name: Active
      cpu_quota:
        description: CPUQuota  allowed cpu seconds
        format: int64
        type: integer
        x-go-name: CPUQuota
      created_at:
        description: CreatedAt creation time
        format: date-time
        type: string
        x-go-name: CreatedAt
      disk_quota:
        description: DiskQuota allowed disk Mbytes
        format: int64
        type: integer
        x-go-name: DiskQuota
      ended_at:
        description: EndedAt ended-at
        format: date-time
        type: string
        x-go-name: EndedAt
      id:
        description: |-
          gorm.Model
          ID defines UUID for primary key
        type: string
        x-go-name: ID
      kind:
        $ref: '#/definitions/Kind'
      organization_id:
        description: OrganizationID defines org
        type: string
        x-go-name: OrganizationID
      period:
        $ref: '#/definitions/Period'
      policy:
        $ref: '#/definitions/Policy'
      price:
        description: Price of subscription in cents
        format: int64
        type: integer
        x-go-name: Price
      remaining_cpu_quota:
        description: RemainingCPUQuota  cpu seconds
        format: int64
        type: integer
        x-go-name: RemainingCPUQuota
      remaining_disk_quota:
        description: RemainingDiskQuota disk Mbytes
        format: int64
        type: integer
        x-go-name: RemainingDiskQuota
      started_at:
        description: StartedAt started-at
        format: date-time
        type: string
        x-go-name: StartedAt
      updated_at:
        description: UpdatedAt update time
        format: date-time
        type: string
        x-go-name: UpdatedAt
      user_id:
        description: UserID defines user
        type: string
        x-go-name: UserID
    type: object
    x-go-package: plexobject.com/formicary/internal/types
...
  UserInvitation:
    description: UserInvitation represents a user session
    properties:
      created_at:
        description: CreatedAt created time
        format: date-time
        type: string
        x-go-name: CreatedAt
      email:
        description: Email defines invitee
        type: string
        x-go-name: Email
      expires_at:
        description: ExpiresAt expiration time
        format: date-time
        type: string
        x-go-name: ExpiresAt
      id:
        description: |-
          gorm.Model
          ID defines UUID for primary key
        type: string
        x-go-name: ID
      invitation_code:
        description: InvitationCode defines code
        type: string
        x-go-name: InvitationCode
      invited_by_user_id:
        description: InvitedByUserID defines foreign key
        type: string
        x-go-name: InvitedByUserID
      org_unit:
        description: OrgUnit defines org-unit
        type: string
        x-go-name: OrgUnit
      organization_id:
        description: OrganizationID defines foreign key
        type: string
        x-go-name: OrganizationID
    type: object
    x-go-package: plexobject.com/formicary/queen/types
...
host: https://formicary.io
info:
  contact:
    email: support@formicary.io
    name: Support
    url: https://formicary.io
  description: |-
    The formicary is a distributed orchestration engine based on `Leader-Follower`, `Pipes-Filter`, `Fork-Join` and `SEDA` design principles for
    executing a directed acyclic graph of tasks, which is also referred as a job workflow. A task represents a unit of work and a job definition is used to specify the task
    dependencies in the graph/workflow including configuration parameters and conditional logic.
  license:
    name: AGPL
    url: https://opensource.org/licenses/AGPL-3.0
  title: Formicary API.
  version: 0.0.1
paths:
...
  /api/jobs/resources:
    get:
      operationId: queryJobResources
      parameters:
      - format: int64
        in: query
        name: page
        type: integer
        x-go-name: Page
      - format: int64
        in: query
        name: page_size
        type: integer
        x-go-name: PageSize
      - description: ResourceType defines type of resource such as Device, CPU, Memory
        in: query
        name: resource_type
        type: string
        x-go-name: ResourceType
      - description: Description of resource
        in: query
        name: description
        type: string
        x-go-name: Description
      - description: Platform can be OS platform or target runtime
        in: query
        name: platform
        type: string
        x-go-name: Platform
      responses:
        "200":
          $ref: '#/responses/jobResourceQueryResponse'
      summary: Queries job resources by criteria such as type, platform, etc.
      tags:
      - job-resources
    post:
      operationId: postJobResource
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/JobResource'
      responses:
        "200":
          $ref: '#/responses/jobResource'
      summary: Adds a job-resource that can be used for managing internal or external
        constraints.
      tags:
      - job-resources
  /api/jobs/resources/{id}:
    get:
      operationId: getJobResource
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/jobResource'
      summary: Finds the job-resource by id.
      tags:
      - job-resources
    put:
      operationId: putJobResource
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/JobResource'
      responses:
        "200":
          $ref: '#/responses/jobResource'
      summary: Updates a job-resource that can be used for managing internal or external
        constraints.
      tags:
      - job-resources
...
  /api/jobs/resources/{id}/configs/{configId}:
    delete:
      operationId: deleteJobResourceConfig
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      - in: path
        name: config_id
        required: true
        type: string
        x-go-name: ConfigID
      responses:
        "200":
          $ref: '#/responses/emptyResponse'
      summary: Deletes the job-resource config by id of job-resource and config-id.
      tags:
      - job-resources
  /api/jobs/resources/{id}/disable:
    post:
      operationId: deleteJobResource
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/emptyResponse'
      summary: Deletes the job-resource by id.
      tags:
      - job-resources
...
  /api/orgs:
    get:
      description: '`This requires admin access`'
      operationId: queryOrganizations
      responses:
        "200":
          $ref: '#/responses/orgQueryResponse'
      summary: Queries organizations by criteria such as org-unit, bundle, etc.
      tags:
      - organizations
    post:
      description: '`This requires admin access`'
      operationId: postOrganization
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/Organization'
      responses:
        "200":
          $ref: '#/responses/orgResponse'
      summary: Creates new organization.
      tags:
      - organizations
  /api/orgs/{id}:
    delete:
      operationId: deleteOrganization
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/emptyResponse'
      summary: Deletes the organization by its id.
      tags:
      - organizations
    get:
      operationId: getOrganization
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/orgResponse'
      summary: Finds the organization by its id.
      tags:
      - organizations
    put:
      operationId: putOrganization
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/Organization'
      responses:
        "200":
          $ref: '#/responses/orgResponse'
      summary: Updates the organization profile.
      tags:
      - organizations
  /api/orgs/{id}/invite:
    post:
      operationId: inviteUser
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/userInvitationResponse'
      summary: Invite a user to the organization.
      tags:
      - organizations
...
  /api/orgs/usage_report:
    post:
      description: |-
        `This requires admin access`
        Shows usage report by organization and user
      operationId: usageReport
      parameters:
      - in: query
        name: from
        type: string
        x-go-name: From
      - description: TO ISO date
        in: query
        name: to
        type: string
        x-go-name: To
      responses:
        "200":
          $ref: '#/responses/usageReportResponse'
      summary: Generates usage report for the organization.
      tags:
      - organizations
...
produces:
- application/json
responses:
...
  emptyResponse:
    description: Empty response body
...
  jobResource:
    description: JobResource represents a virtual resource, which can be used to implement
      mutex/semaphores for a job.
    schema:
      $ref: '#/definitions/JobResource'
...
  jobResourceQueryResponse:
    description: Paginated results of jobResources matching query
    schema:
      properties:
        Page:
          format: int64
          type: integer
        PageSize:
          format: int64
          type: integer
        Records:
          items:
            $ref: '#/definitions/JobResource'
          type: array
        TotalPages:
          format: int64
          type: integer
        TotalRecords:
          format: int64
          type: integer
      type: object
...
  orgQueryResponse:
    description: Paginated results of orgs matching query
    schema:
      properties:
        Page:
          format: int64
          type: integer
        PageSize:
          format: int64
          type: integer
        Records:
          items:
            $ref: '#/definitions/Organization'
          type: array
        TotalPages:
          format: int64
          type: integer
        TotalRecords:
          format: int64
          type: integer
      type: object
  orgResponse:
    description: Org defines user request to process a job, which is saved in the
      database as PENDING and is then scheduled for job execution.
    schema:
      $ref: '#/definitions/Organization'
...
  usageReportResponse:
    description: Usage Report
    schema:
      items:
        $ref: '#/definitions/CombinedResourceUsage'
      type: array
  userInvitationResponse:
    description: User invitation
    schema:
      $ref: '#/definitions/UserInvitation'
...
schemes:
- http
- https
security:
- api_key: []
securityDefinitions:
  api_key:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...